Refactoring: Fixing Reference problems caused by the last commit.
