+++ DroidPlanner/src/com/droidplanner/file/DirectoryPath.java
@@ -1,4 +1,4 @@
-package com.droidplanner.helpers.file;
+package com.droidplanner.file;
 
 import android.os.Environment;
