+++ DroidPlanner/src/com/droidplanner/file/FileManager.java
@@ -1,4 +1,4 @@
-package com.droidplanner.helpers.file;
+package com.droidplanner.file;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
