+++ DroidPlanner/src/com/droidplanner/helpers/LocalMapTileProvider.java
@@ -6,7 +6,7 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.Locale;
 
-import com.droidplanner.helpers.file.DirectoryPath;
+import com.droidplanner.file.DirectoryPath;
 import com.google.android.gms.maps.model.Tile;
 import com.google.android.gms.maps.model.TileProvider;
